Created a notification controller and service.
The controller can be used to get all the notification of an user to the frontend.
Each notification consists of
the userId (the user to whom the notification belongs)
the text (what should the user be informed of)
read status (has the user already read the notification or not?) (0; not read, 1: read)

The notification can be used to create new notification entries in the database.
Currently this is only used for the frontend to be able to show the notifications.
Later this should be somewhat the control place to send out emails and push notifications.